all test covered & passed...more bitter leaf tea☕️
this commit completes all test in the js/ts sdk, therefore it can be marked
as complete with the js/ts sdk at commit a9466fa0061473160c4369b09cc6fc1766df0b99.

Next task:
 - aDD  a little README.md and introduce it to the Meteora team.
 - try as much to re-read and do little cleaning (optimisation) where possible.
 - buying more fresh bitter-leaf tea.

+++ types/types.go
@@ -445,7 +445,7 @@ type FeeSchedulerParams struct {
 type RateLimiterParams struct {
 	BaseFeeBps         uint64
 	FeeIncrementBps    uint16
-	ReferenceAmount    uint64
+	ReferenceAmount    float64
 	MaxLimiterDuration uint64
 }
